Parse `pylint_options` depending on newlines (#170)
Clarify when single space-separated line is accepted

If it not accepted: Remove the mention of single line in the description
of the input, leaving only "multi-line". And always parse as if it is a
multi-line input.
If it is accepted: Update the parsing to check for newline characters
before actually parsing the input. Also add "(space-separated)" to
describe a multi-valued single-line input.

A new section "General information" has been added in the documentation
under the overview of the workflows.
A section about single vs multi-line values have been added, including
examples.
This section has then been referenced throughout the documentation for
specific input parameters that mention "single og multi-line values".