Create ignore rules based on requirement specifiers
Depending on each dependency's set of specifiers, a list of on-the-fly
ignore rules are created and added to the otherwise user-provided ignore
rules.
This is a way of using the existing implementation of "ignore rules" to
more intelligently suggest an updated set of version specifiers for a
dependency based on the latest version retrieved from 'pip index
versions'.