Introduce the `system_dependencies` input (#328)
Adds the system_dependencies input to all workflows
that install the local Python package. 
his is to install Ubuntu APT package prior to installing
the Python package.

Where the system_dependencies was already defined,
it has now been elevated to a general input that will work
for other jobs in the same workflow as well.

Update documentation accordingly.